[CoreComponent] Introduce IdentifiableUser abstact class
