+++ src/Sylius/Component/Core/Model/AdminUser.php
@@ -11,74 +11,9 @@
 
 namespace Sylius\Component\Core\Model;
 
-use Doctrine\Common\Collections\ArrayCollection;
-use Sylius\Component\Rbac\Model\RoleInterface;
-use Sylius\Component\User\Model\User;
-
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
  */
-class AdminUser extends User implements AdminUserInterface
+class AdminUser extends IdentifiableUser implements AdminUserInterface
 {
-    /**
-     * @var ArrayCollection
-     */
-    protected $authorizationRoles;
-
-    public function __construct()
-    {
-        parent::__construct();
-
-        $this->authorizationRoles = new ArrayCollection();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getAuthorizationRoles()
-    {
-        return $this->authorizationRoles;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addAuthorizationRole(RoleInterface $role)
-    {
-        if (!$this->hasAuthorizationRole($role)) {
-            $this->authorizationRoles->add($role);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function removeAuthorizationRole(RoleInterface $role)
-    {
-        if ($this->hasAuthorizationRole($role)) {
-            $this->authorizationRoles->removeElement($role);
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function hasAuthorizationRole(RoleInterface $role)
-    {
-        return $this->authorizationRoles->contains($role);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getRoles()
-    {
-        $roles = parent::getRoles();
-
-        foreach ($this->getAuthorizationRoles() as $role) {
-            $roles = array_merge($roles, $role->getSecurityRoles());
-        }
-
-        return $roles;
-    }
 }
