+++ src/Sylius/Component/Core/Model/AdminUserInterface.php
@@ -11,29 +11,9 @@
 
 namespace Sylius\Component\Core\Model;
 
-use Sylius\Component\Rbac\Model\IdentityInterface;
-use Sylius\Component\Rbac\Model\RoleInterface;
-use Sylius\Component\User\Model\UserInterface as BaseUserInterface;
-
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
  */
-interface AdminUserInterface extends BaseUserInterface, IdentityInterface
+interface AdminUserInterface extends IdentifiableUserInterface
 {
-    /**
-     * @param RoleInterface $role
-     */
-    public function addAuthorizationRole(RoleInterface $role);
-
-    /**
-     * @param RoleInterface $role
-     */
-    public function removeAuthorizationRole(RoleInterface $role);
-
-    /**
-     * @param RoleInterface $role
-     *
-     * @return bool
-     */
-    public function hasAuthorizationRole(RoleInterface $role);
 }
