+++ src/Sylius/Component/Core/Model/ShopUserInterface.php
@@ -12,29 +12,10 @@
 namespace Sylius\Component\Core\Model;
 
 use Sylius\Component\Customer\Model\CustomerAwareInterface;
-use Sylius\Component\Rbac\Model\IdentityInterface;
-use Sylius\Component\Rbac\Model\RoleInterface;
-use Sylius\Component\User\Model\UserInterface as BaseUserInterface;
 
 /**
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-interface ShopUserInterface extends BaseUserInterface, IdentityInterface, CustomerAwareInterface
+interface ShopUserInterface extends IdentifiableUserInterface, CustomerAwareInterface
 {
-    /**
-     * @param RoleInterface $role
-     */
-    public function addAuthorizationRole(RoleInterface $role);
-
-    /**
-     * @param RoleInterface $role
-     */
-    public function removeAuthorizationRole(RoleInterface $role);
-
-    /**
-     * @param RoleInterface $role
-     *
-     * @return bool
-     */
-    public function hasAuthorizationRole(RoleInterface $role);
 }
